mmc-utils: Reduce scope of nargs for RPMB commands
This commit fixes the loose definition of the `nargs` member in the
`struct Command` for RPMB commands. The `nargs` member specifies the
number of arguments required for each command, and this change ensures
that the correct number of arguments is enforced for each RPMB command.

This reduction in the scope of the `nargs` variable helps to prevent
incorrect usage of the RPMB commands by ensuring that the correct number
of arguments is provided.

Signed-off-by: Avri Altman <avri.altman@sandisk.com>
Link: https://lore.kernel.org/r/20250407082833.108616-3-avri.altman@sandisk.com
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>