mmc-utils: Add byte offset comments to rpmb_frame structure
Added comments to the `rpmb_frame` structure to document the byte
offsets of each field within the 512-byte packed frame. The
specification defines the frame format in terms of byte offsets, and
these comments make it easier to align the code with the spec and
improve readability.

No functional changes were made.

Signed-off-by: Avri Altman <avri.altman@sandisk.com>
Link: https://lore.kernel.org/r/20250407082833.108616-5-avri.altman@sandisk.com
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>